Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> I see one extra file for DEB vs RPM.   Is the "dbgsym" file needed for debian distributions but not for redhat?  
> 
> clott@ubuntu-cl778h:/tmp/export$ ls -l *.deb
> -rw-r--r-- 1 root root  6592 Jul 31 12:26 mdclog_0.0.3-1_amd64.deb
> -rw-r--r-- 1 root root 14958 Jul 31 12:26 mdclog-dbgsym_0.0.3-1_amd64.deb
> -rw-r--r-- 1 root root  3736 Jul 31 12:26 mdclog-dev_0.0.3-1_amd64.deb
> clott@ubuntu-cl778h:/tmp/export$ ls -l *.rpm
> -rw-r--r-- 1 root root 24791 Jul 31 12:26 mdclog-0.0.3-1.x86_64.rpm
> -rw-r--r-- 1 root root  9059 Jul 31 12:26 mdclog-devel-0.0.3-1.x86_64.rpm

Actually, for rpm it depends how the build environment is configured. It seems to be disabled in that Docker image.

Debian build produces dbg packages by default.

I don't know if we need those or not. I can try to disable them if you like or you can ignore them when pushing them to LF.

Patch-set: 6
Label: Code-Review=0