Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> I have one minor question. I see your configure.ac script sets the version to 0.0.3, yet during the build a destination directory is created with a "-1" suffix, and the deb/rpm files are written with version suffix 0.0.3-1.  Is the "-1" suffix important?

Maybe not in our case when the version would always change. In real Linux distros where the packaging is separated from the sources the release number can increase while version stays the same.
But the tools always add that so we just have to live with it.

Patch-set: 7